Fix DateRangePicker build
